refactoring stitcher package
