+++ code/src/terastitcher_module/__init__.py
@@ -2,6 +2,3 @@
 Init file
 """
 __version__ = "1.0.0"
-
-# flake8: noqa: F401
-from terastitcher_module import terastitcher
