+++ code/src/terastitcher_module/utils/utils.py
