+++ code/src/terastitcher_module/zarr_converter/__init__.py
@@ -1,0 +1,4 @@
+"""
+Init file to the zarr converter package
+"""
+# flake8: noqa: F401
