+++ code/src/terastitcher_module/zarr_converter/zarr_converter_params.py
