--- code/src/zarr_converter/__init__.py
@@ -1,6 +1,0 @@
-"""
-Init file to the zarr converter package
-"""
-# flake8: noqa: F401
-from zarr_converter.zarr_converter import ZarrConverter
-from zarr_converter.zarr_converter_params import OmeZarrParams
