+++ metadata/metadata.yml
@@ -1,5 +1,5 @@
 metadata_version: 1
-name: TeraStitcher Module
+name: SmartSPIM Stitcher
 description: A tool for fast automatic 3D-stitching of teravoxel-sized microscopy images using cloud computing resources and TeraStitcher.
 authors:
 - name: David Feng
